requirements.txt
-i https://pypi.org/simple
alembic==1.13.0; python_version >= '3.8'
aniso8601==9.0.1
bcrypt==4.1.1; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
cachelib==0.10.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
faker==21.0.0; python_version >= '3.8'
flask==3.0.0; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5; python_version >= '3.6'
flask-restful==0.3.10
flask-session==0.5.0; python_version >= '3.7'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.9.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
